CDDSO-365: Adding capability to handle generic MIP tables for CMIP6Plus (from v2.5_release) (#192)
CDDSO-365: Changes to allow for use of generic MIP tables for CMIP6Plus:

This change is a set of cherry picks from pull request #191 

* Corrected grids for CMIP6Plus Ocean MIP tables
* renamed cimp_plus_attributes to cmip_plus_attributes
* A change deep in the bowels of MIP Convert. The code in cmor_outputter
is not well understood, but this change is needed as the generic_levels entry
in the tables is now a list rather than a string. Previously an empty
string was being looked for within a string, which returned True. With the
change to a list this returned False. This code could do with refactoring.
* set up submission code for CMIP6Plus transfers including use of sub_experiment_id
* Correct grids and mappings
* Add APmon/hur mapping (on plev19, so separate table needed)
* Changes needed to support use of arbitrary root proc directory in submission code
* remove un-necessary exclusion from facet matching in submission conde
* Fix send_to_mass tests broken by changes made to command line for submission code